requirements.txt
Django==3.2
psycopg2==2.9.9
faker==23.3.0
django-encrypted-model-fields==0.6.5
django-silk==4.3.0
djangorestframework==3.14.0